test: refactor the directory structure of test files
